+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/inline_adapter_consumer.php
@@ -170,6 +170,11 @@ class Symfony_DI_PhpDumper_Test_Inline_Adapter_Consumer extends Container
         throw new ParameterNotFoundException($name);
     }
 
+    public function isParameterDynamic(string $name): bool
+    {
+        return isset($this->loadedDynamicParameters[$name]);
+    }
+
     protected function getDefaultParameters(): array
     {
         return [