+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_default_env.php
@@ -92,6 +92,11 @@ class Symfony_DI_PhpDumper_Test_DefaultParameters extends Container
         return $this->dynamicParameters[$name] = $value;
     }
 
+    public function isParameterDynamic(string $name): bool
+    {
+        return isset($this->loadedDynamicParameters[$name]);
+    }
+
     protected function getDefaultParameters(): array
     {
         return [