+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_json_env.php
@@ -90,6 +90,11 @@ class Symfony_DI_PhpDumper_Test_JsonParameters extends Container
         return $this->dynamicParameters[$name] = $value;
     }
 
+    public function isParameterDynamic(string $name): bool
+    {
+        return isset($this->loadedDynamicParameters[$name]);
+    }
+
     protected function getDefaultParameters(): array
     {
         return [