+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_url_env.php
@@ -88,6 +88,11 @@ class Symfony_DI_PhpDumper_Test_UrlParameters extends Container
         return $this->dynamicParameters[$name] = $value;
     }
 
+    public function isParameterDynamic(string $name): bool
+    {
+        return isset($this->loadedDynamicParameters[$name]);
+    }
+
     protected function getDefaultParameters(): array
     {
         return [